move mapstyle
